Fix repeating appointments, Fix: #169
